deleted /api from comments route and moved models types to types folder
