+++ backend/src/types/reaction-model.ts
@@ -1,0 +1,7 @@
+import { ObjectId } from 'mongoose';
+
+export interface IStudent {
+  _id: ObjectId;
+  name: string;
+  email: string;
+}
